5758 - Remove validateCountries and use validateCycleCountries instead
