+++ src/tools/validations/validateDataSources.ts
@@ -1,88 +1,40 @@
 import '../scriptInit'
 
-import { Country } from 'meta/area/country'
 import { CountryIso } from 'meta/area/countryIso'
-import { Assessment } from 'meta/assessment/assessment'
-import { Cycle } from 'meta/assessment/cycle'
-import {
-  CommentableDescription,
-  CommentableDescriptionName,
-  DescriptionCountryValues,
-} from 'meta/assessment/descriptionValue'
+import { Promises } from 'utils/promises'
 import { ToolsUtils } from 'tools/utils/toolsUtils'
 
 import { AreaController } from 'server/controller/area'
-import { AssessmentController } from 'server/controller/assessment'
 import { BaseProtocol, DB } from 'server/db/db'
 import { DescriptionRepository } from 'server/db/repository/assessmentCycle/descriptions'
-import { DataValidationService } from 'server/service/dataValidation'
 import { Logger } from 'server/utils/logger'
 
 import { Failures } from './common/failures'
+import { getAssessmentCycles } from './common/getAssessmentCycles'
 import { Failure } from './common/types'
+import { validateCycleCountries } from './common/validateCycleCountries'
+import { validateCountryDataSources } from './dataSources/validateCountryDataSources'
 
 const toolName = 'validateDataSources'
 
-type CountryProps = {
-  assessment: Assessment
-  country: Country
-  cycle: Cycle
-  descriptionsByCountry: DescriptionCountryValues
-}
-
-const _validateCountry = async (props: CountryProps): Promise<void> => {
-  const { assessment, country, cycle, descriptionsByCountry } = props
-  const { countryIso } = country
-
-  const descriptions = Object.entries(descriptionsByCountry[countryIso] ?? {}).flatMap<
-    Omit<CommentableDescription, 'id'>
-  >(([sectionName, sectionValues]) =>
-    Object.entries(sectionValues).map<Omit<CommentableDescription, 'id'>>(([name, value]) => ({
-      countryIso,
-      name: name as CommentableDescriptionName,
-      sectionName,
-      value,
-    }))
-  )
-
-  await DataValidationService.removeDescriptionValidations({ assessment, countryIso, cycle })
-
-  await DataValidationService.validateDataSources({ assessment, country, cycle, descriptions, notifyClients: false })
-}
-
 export const validateDataSources = async (client: BaseProtocol = DB): Promise<Array<Failure>> => {
   const failures: Array<Failure> = []
-  const assessments = await AssessmentController.getAll({ metaCache: true }, client)
-  const assessmentCycles = assessments.flatMap((assessment) =>
-    assessment.cycles.map((cycle) => ({ assessment, cycle }))
-  )
+  const assessmentCycles = await getAssessmentCycles(client)
 
-  await Promise.all(
-    assessmentCycles.map(async ({ assessment, cycle }) => {
-      const assessmentName = assessment.props.name
-      const { name: cycleName } = cycle
-      Logger.info(`${toolName}: ${assessmentName}/${cycleName}`)
+  await Promises.each(assessmentCycles, async ({ assessment, cycle }) => {
+    Logger.info(`${toolName}: ${assessment.props.name}/${cycle.name}`)
 
-      const countries = await AreaController.getCountries({ assessment, cycle }, client)
-      const descriptionsByCountry = await DescriptionRepository.getValues(
-        { assessment, countryISOs: countries.map<CountryIso>(({ countryIso }) => countryIso), cycle },
-        client
-      )
+    const countries = await AreaController.getCountries({ assessment, cycle }, client)
+    const descriptionsByCountry = await DescriptionRepository.getValues(
+      { assessment, countryISOs: countries.map<CountryIso>(({ countryIso }) => countryIso), cycle },
+      client
+    )
 
-      await Promise.all(
-        countries.map(async (country) => {
-          const { countryIso } = country
-
-          try {
-            await _validateCountry({ assessment, country, cycle, descriptionsByCountry })
-          } catch (error) {
-            Logger.error(`${toolName} failed for ${assessmentName}/${cycleName}/${countryIso}`)
-            failures.push({ assessmentName, countryIso, cycleName, error })
-          }
-        })
-      )
-    })
-  )
+    const cycleFailures = await validateCycleCountries({ assessment, countries, cycle, toolName }, (country) =>
+      validateCountryDataSources({ assessment, country, cycle, descriptionsByCountry })
+    )
+    failures.push(...cycleFailures)
+  })
 
   return failures
 }
